Working in user CRUD

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -29,6 +29,7 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('users');
+        if(Schema::hasTable('users'))
+            Schema::drop('users');
     }
 }
