+++ database/migrations/2014_10_12_100000_create_password_resets_table.php
@@ -26,6 +26,7 @@ class CreatePasswordResetsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('password_resets');
+        if(Schema::hasTable('password_resets'))
+            Schema::drop('password_resets');
     }
 }
